+++ src/changelog/updateChangelog.ts
@@ -1,3 +1,4 @@
+import { commitChanges } from '../release/commitChanges';
 import fs from 'fs';
 
 /**
@@ -5,25 +6,20 @@ import fs from 'fs';
  *
  * @param {string} markdown String to append
  */
-export const updateChangelog = (markdown: string) => {
+export const updateChangelog = async (markdown: string, packageVersion: string) => {
   const changelogPath = `${process.cwd()}/CHANGELOG.md`;
-  fs.readFile(changelogPath, 'utf8', (err, data) => {
-    if (err) {
-      console.log(err);
-    } else {
-      if (data !== null) {
-        // Split contents by line breaks and insert the newest markdown right after CHANGELOG header
-        const oldSections = data.split('\n');
-        oldSections.splice(1, 0, markdown);
-        const newSections = oldSections.join('\n');
-        fs.writeFile(changelogPath, newSections, (err) => {
-          if (err) {
-            console.log(err);
-          } else {
-            console.log('Data added to CHANGELOG file');
-          }
-        });
-      }
-    }
-  });
+  const oldChangelog: string = fs.readFileSync(changelogPath, {encoding: 'utf8'});
+
+  if (oldChangelog !== null) {
+    // Split contents by line breaks and insert the newest markdown right after CHANGELOG header
+    const oldSections = oldChangelog.split('\n');
+    oldSections.splice(1, 0, markdown);
+    const newSections = oldSections.join('\n');
+    
+    fs.writeFileSync(changelogPath, newSections);
+  } else {
+    console.error('CHANGELOG.md is empty...please initialize by adding "--first-release" as an argument')
+  }
+
+  commitChanges(packageVersion);
 };