+++ src/commit/ParsedMessages.d.ts
@@ -1,0 +1,9 @@
+import { RegexCaptureGroups } from "./RegexCaptureGroups";
+
+/**
+ *  Interface declaring how ```parseCommitMsg``` will return the data from git logs when preparing to modify ```CHANGELOG.md```.
+ */
+export interface ParsedMessages {
+    shortHash: string;
+    regexGroups: RegexCaptureGroups;
+}