+++ src/commit/getCommitDate.ts
@@ -5,7 +5,7 @@ import { execSync } from 'child_process';
  *
  * @returns {string} Date of the most recent commit
  */
-export const getCommitDate = (): string => {
+export const getCommitDate = async (): Promise<string> => {
   const command: string = 'git log -1 --pretty="format:%cd" --date=format:"%Y-%m-%d"';
   return execSync(command).toString().trim();
 };