+++ src/index.ts
@@ -1,4 +1,8 @@
 import { createChangelogMessage } from "./changelog/createChangelogMessage";
 
 // Entry point for conventionalLogs
-createChangelogMessage()
+try {
+    createChangelogMessage()
+} catch (err) {
+    console.error(err)
+}